Unit tests: Drop node 6 and 7 for 9, split tests
Move tests to separate files.
