+++ .travis.yml
@@ -1,8 +1,7 @@
 language: node_js
 node_js:
-  - "6"
-  - "7"
   - "8"
+  - "9"
 after_script:
   - sonar-scanner -Dsonar.login=$SONAR_TOKEN
 after_success:
