+++ tests/spinner.js
@@ -1,0 +1,51 @@
+/*
+Unit tests for the visualization spinner fragment.
+
+Copyright 2017 ICTU
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+/* jshint node: true, mocha: true */
+'use strict';
+
+const assert = require('chai').assert,
+      setupPage = require('./index');
+
+describe('Spinner', () => {
+    it('Creates the spinner', (done) => {
+        const { window, d3, spinner } = setupPage('<div id="loader_container"></div>', done);
+        const loadingSpinner = new spinner();
+        loadingSpinner.start();
+        assert.isFalse(d3.select("svg#loader").empty(), 'Spinner has SVG');
+        done();
+    });
+    it('Creates at most one spinner with same ID', (done) => {
+        const { window, d3, spinner } = setupPage('<div id="loader_container"></div>', done);
+        const loadingSpinner = new spinner();
+        loadingSpinner.start();
+        loadingSpinner.start();
+        assert.equal(d3.selectAll("svg#loader").size(), 1);
+        done();
+    });
+    it('Removes the spinner and its events', (done) => {
+        const { window, d3, spinner } = setupPage('<div id="loader_container"></div>', done);
+        const duration = 25;
+        const loadingSpinner = new spinner({ duration: duration });
+        loadingSpinner.start();
+        setTimeout(function() {
+            loadingSpinner.stop();
+            assert.isTrue(d3.select("svg#loader").empty(), 'Spinner is removed');
+            setTimeout(done, duration*2);
+        }, duration);
+    });
+});
